+++ NEWS.md
@@ -5,6 +5,8 @@ User-visible changes worth mentioning.
 ## master
 
 - [#907] Fix lookup for matching tokens in certain edge-cases
+- [#992] Add API option to use Doorkeeper without management views for API only
+  Rails applications (`api_only`)
 - [#1045] Validate redirect_uri as the native URI when making authorization code requests
 
 ## 4.3.1